Fix users endpoint and upgrade jest
The users list endpoint was failing to parse the response because
crudOperations couldn't handle URL without a trailing slash

I added a test to work through that, which showed a mocking reset
test issue that could be covered easily by a new jest version.
So I upgraded jest while I was here.